+++ etc/build-help-main-1.sh
@@ -1,0 +1,9 @@
+
+src=$1
+
+sed -e '/HELP_CMD_INDEX/,$d' < $src
+
+cat var/help-cmd-list.txt | sed 's/^/        /'
+
+sed -n -e '/HELP_CMD_INDEX/,$p' < $src | tail -n+2
+